Restructure midi interface
