+++ src/core/core.cpp
@@ -1,4 +1,5 @@
 #include "core.hpp"
+#include "MidiInterface.hpp"
 
 
 void init(rack::Plugin *plugin) {
@@ -8,6 +9,7 @@ void init(rack::Plugin *plugin) {
 	createModel<AudioInterfaceWidget>(plugin, "AudioInterface", "Audio Interface");
 	createModel<MidiToCVWidget>(plugin, "MIDIToCVInterface", "MIDI-to-CV Interface");
 	createModel<MIDICCToCVWidget>(plugin, "MIDICCToCVInterface", "MIDI CC-to-CV Interface");
+	createModel<MIDIClockToCVWidget>(plugin, "MIDIClockToCVInterface", "MIDI Clock-to-CV Interface");
 	// createModel<BridgeWidget>(plugin, "Bridge", "Bridge");
 	createModel<BlankWidget>(plugin, "Blank", "Blank");
 }
